fixed timeouts

+++ hw6/hw6.h
@@ -14,7 +14,7 @@ struct hw6_hdr {
 #define INIT_TO 1000 //msec, initial estimated rtt is 1 sec
 #define MAX_PACKET 1400
 #define MAX_SEGMENT (MAX_PACKET-sizeof(struct hw6_hdr))
-#define HDR_SZ (sizeof(struct hw6_hdr))
+#define HDR_SZ ((int)sizeof(struct hw6_hdr))
 #define diff(x,y) ((x)< (y) ? (y)-(x) : (x)-(y)) 
 
 typedef struct hw6_hdr* hdr_ptr;
